+++ resources/views/permissions/index.blade.php
@@ -7,13 +7,11 @@
 			<div class="pull-left">
 				<h2>Permission Management</h2>
 			</div>
-			<div class="pull-right">
-
+			@can('permission-create')
+				<div class="pull-right">
 					<a class="btn btn-success" href="{{ route('permissions.create') }}"> Create New Permission</a>
-
-
-
-			</div>
+				</div>
+			@endcan
 			<br>
 		</div>
 	</div>
@@ -37,13 +35,17 @@
 				<td>{{ ++$i }}</td>
 				<td>{{ $permission->name }}</td>
 				<td>
-					<a class="btn btn-info" href="{{ route('permissions.show',$permission->id) }}">Show</a>
-
+					@can('permission-show')
+						<a class="btn btn-info" href="{{ route('permissions.show',$permission->id) }}">Show</a>
+					@endcan
+					@can('permission-edit')
 						<a class="btn btn-primary" href="{{ route('permissions.edit',$permission->id) }}">Edit</a>
-
+					@endcan
+					@can('permission-delete')
 						{!! Form::open(['method' => 'DELETE','route' => ['permissions.destroy', $permission->id],'style'=>'display:inline']) !!}
 						{!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
 						{!! Form::close() !!}
+					@endcan
 
 				</td>
 			</tr>
